feat: add TextSelectableStateMixin
