+++ frontend/appflowy_flutter/packages/appflowy_editor/lib/src/blocks/quote_block/quote_block.dart
@@ -1,7 +1,7 @@
 import 'package:appflowy_editor/appflowy_editor.dart';
+import 'package:appflowy_editor/src/blocks/base_component/selectable/text_selectable_state_mixin.dart';
 import 'package:appflowy_editor/src/blocks/text_block/text_block_with_icon.dart';
 import 'package:appflowy_editor/src/infra/flowy_svg.dart';
-import 'package:appflowy_editor/src/render/selection/v2/selectable_v2.dart';
 import 'package:flutter/material.dart';
 
 class QuoteBlock extends StatefulWidget {
@@ -16,22 +16,7 @@ class QuoteBlock extends StatefulWidget {
   State<QuoteBlock> createState() => _QuoteBlockState();
 }
 
-class _QuoteBlockState extends State<QuoteBlock>
-    implements SelectableState<QuoteBlock> {
-  final GlobalKey textBlockKey = GlobalKey();
-
-  @override
-  Position getPositionInOffset(Offset offset) {
-    return (textBlockKey.currentState as TextBlockState)
-        .getPositionInOffset(offset);
-  }
-
-  @override
-  Future<void> setSelectionV2(Selection? selection) {
-    return (textBlockKey.currentState as TextBlockState)
-        .setSelectionV2(selection);
-  }
-
+class _QuoteBlockState extends State<QuoteBlock> with TextSelectableStateMixin {
   @override
   Widget build(BuildContext context) {
     final node = widget.node;
