+++ frontend/appflowy_flutter/packages/appflowy_editor/lib/src/render/selection/v2/selectable_v2.dart
@@ -2,7 +2,7 @@ import 'package:appflowy_editor/src/core/location/position.dart';
 import 'package:appflowy_editor/src/core/location/selection.dart';
 import 'package:flutter/material.dart';
 
-abstract class SelectableState<T extends StatefulWidget> extends State<T> {
+abstract class SelectableState {
   Future<void> setSelectionV2(Selection? selection);
 
   Position getPositionInOffset(Offset offset);
